Move equalHeightBlob to logic dir Rename logic.py feature_discovery.py
